I have separated into differents components
